python_gui/requirements.txt
PyQt6>=6.5.0
pyserial>=3.5
pyqtgraph>=0.13.0
numpy>=1.21.0